Refactor project structure by moving main and renderer files into a 'src' directory; update references in HTML and package.json accordingly. Remove obsolete main.js, renderer.js, setup.js, payload-viewer.js, and related backend files to streamline the codebase.
