+++ index.html
@@ -14,7 +14,7 @@
     <link rel="stylesheet" href='codemirror/codemirror.css' rel='stylesheet'>
     <title>Kafka Safe Stream</title>
     <link rel="stylesheet" href="styles.css">
-    <script src="./renderer.js"></script>
+    <script src="./src/renderer/renderer.js"></script>
     <script src='codemirror/codemirror.min.js'></script>
     <script src='codemirror/javascript.min.js'></script>
     <script src='codemirror/xml.min.js'></script>
