+++ payload-viewer.html
@@ -15,6 +15,6 @@
     <script src="codemirror/codemirror.min.js"></script>
     <script src="codemirror/javascript.min.js"></script>
     <script src="codemirror/xml.min.js"></script>
-    <script src="payload-viewer.js"></script>
+    <script src="src/renderer/payload-viewer.js"></script>
 </body>
 </html>
