+++ setup.html
@@ -54,6 +54,6 @@
         </main>
       </div>
     </div>
-    <script src="setup.js"></script>
+    <script src="src/renderer/setup.js"></script>
   </body>
 </html>
