+++ src/backend/kafka.js
