+++ src/backend/kafkaConnection.js
