+++ src/renderer/payload-viewer.js
